scripts/install-halon-services: use eth1 for HALOND_LISTEN
If cmu has several interfaces - the old code would pick up
only the last one and take its IP address for the HALON_LISTEN
parameter. But on the vagrant setup, for example, this might
be the bridge interface and the wrong one.

Now we always take eth1 - the same way when we generate the
facts file from scripts/h0.