HALON-893: fix RC restarts on a slow/busy storage
There RS lease timeouts and RC restarts may happen on a
slow or busy storage.  The issue was observed during the
t_stop-reverts-sdev-states system test on GitLab when we write
about 4GB of data to the cluster in two parallel client threads
that makes the storage very busy when flushing the large host
caches to HDD.

Now we set the rs-lease timeout to 30 secs (default is 4 secs)
via HALOND_STATION_OPTIONS parameter in /etc/sysconfig/halond.
It solves the problem on the GitLab's VM.