hctl: fix error when HALOND_STATION_OPTIONS is unset
This is the addon to commit 8d71b60 where we started using
HALOND_STATION_OPTIONS parameter. hctl script uses `-u' option
so it complained like this:

/usr/bin/hctl: line 31: HALOND_STATION_OPTIONS: unbound variable

if the parameter was unset at /etc/sysconfig/halond.

Now we explicitly set this variable to default value at hctl
script if it is unset.