HALON-900: fix multi-nodes TS cluster bootstrap regression
The same lease timeout which we increased in the commit 8d71b602
is also used by the ambassador to ping the replicas and determine
the leader. So the leader determination increased as well along
with the RC startup time. As a result, the satellites startup
timed out during the bootstrap.

Now we just hard-code the initial ambassador timeout to 1 sec. It
seems to be enough to quickly and reliably determine the leader
during the bootstrap and also break the dependency of this
determination process on the lease timeout configuration.