scripts/h0: fix cmd_install on vagrant singlenode
On vagrant setup with bridge interface the cmd_install was
selecting the wrong interface and, as result, wrong IP address
for HALOND_LISTEN parameter on the singlenode configuration.

Now we just select eth1 interface (if present) which is the
default interface on vagrant setup. Otherwise, we allow
scripts/install-halon-services to select the interface.