network: fix rare data corruption on sendMany
The problem was that at withIOVec routine the elements of
the ByteString array could be carbage collected already
by the time c_writev system call use them. As result,
the data could be sent corrupted.

Now we make sure that the ByteStrings are kept in memory
by referencing them after the c_writev call also.