COSTOR-894: fix notification failures during cluster start/stop
The default 2 seconds delay for the notifications aggregation
introduced at commit 20a96f16 was too small to capture all the
notifications, so the CEP engine suffered again from overloading
with processing too many notification Acks. It was noticed on
as large as 7 SSUs configurations already.

Now we improved the aggregation algorithm by introducing an
additional idle timer for the new notifications to be added
into the aggregation (we reused old parameter name for this
`--notification-aggr-delay`, 5 seconds by default). And also by
increasing the existing maximum time for the aggregation buildup
from 2 to 20 seconds (we renamed old parameter name for this to
`--notification-aggr-max-delay`). Thus we don't wait too long
when there are few notifications to be sent, from one side. And,
from another side, we wait long enough to accumulate all the
notifications when there are many of them.