HALON-911: fix processes restart after node failure
It was possible that the node's processes would get stuck
in a failed state with 'node failure' status on cluster startup
sometimes. It happened because the node::process::start rule
instance was not finishing until all the cluster processes are
started up (including the client ones). As result, the attempt
to restart the processes on the failed and restored node was
failing because the node::process::start rule instance was
'already running'.

Now we finish the rule instance as soon as there are no more
processes left to start on the node or some previously started
processes on the node got failed already (due to the node failure,
for example).